Move pdffit2 to a dedicated subfolder
