+++ easydiffraction/Interfaces/__init__.py
@@ -3,18 +3,18 @@
 # © 2021-2024 Contributors to the EasyDiffraction project <https://github.com/easyscience/EasyDiffraction
 
 try:
-    from easydiffraction.calculators.cryspy.wrapper import CryspyBase  # noqa: F401
+    from easydiffraction.calculators.cryspy.wrapper import CryspyWrapper  # noqa: F401
 except ImportError:
     print('Warning: CrysPy is not installed')
 
 # Temporarily disabling the PyCrysFML interface
 # try:
-#     from easydiffraction.calculators.pycrysfml.wrapper import Pycrysfml  # noqa: F401
+#     from easydiffraction.calculators.pycrysfml.wrapper import PycrysfmlWrapper  # noqa: F401
 # except ImportError:
 #     print('Warning: CrysFML is not installed')
 
 try:
-    from easydiffraction.Interfaces.pdffit2 import Pdffit2  # noqa: F401
+    from easydiffraction.calculators.pdffit2.wrapper import Pdffit2Wrapper  # noqa: F401
 except ImportError:
     pass
     #print('Warning: PdfFit2 is not installed')
